% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\docType{data}
\name{lisa}
\alias{lisa}
\title{List of all available palettes}
\format{An object of class \code{list} of length 128.}
\usage{
lisa
}
\description{
A list containing all available palettes (128) and their attributes (artist
name, name of work).
}
\details{
Source: \url{http://colorlisa.com/}
}
\keyword{datasets}
