% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\docType{data}
\name{work}
\alias{work}
\title{Name of work and artist}
\format{A data frame with 128 rows and 2 variables:
\describe{
  \item{work}{name of work}
  \item{artist}{name of artist}
}

Source: \url{http://colorlisa.com/}}
\usage{
work
}
\description{
A dataset containing the artist name and the name of the work which the palette
comes from. These values are also recorded as \code{attributes} for each palette.
}
\keyword{datasets}
