% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\docType{data}
\name{artwork}
\alias{artwork}
\title{Artist names, palette names, artwork names}
\format{A data frame with 128 rows and 3 variables:
\describe{
  \item{artist}{name of artist}
  \item{palette}{name of the palette}
  \item{work}{name of artwork}
}

Source: \url{http://colorlisa.com/}}
\usage{
artwork
}
\description{
A dataset containing the name of the artist, palette, and artwork.
}
\keyword{datasets}
