% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\docType{data}
\name{artwork}
\alias{artwork}
\title{Artist, palettes, and artwork names}
\format{
A data frame with 128 rows and 3 variables:
\describe{
\item{author}{name of artist}
\item{work}{name of artwork}
\item{palette}{name of the palette}
\item{colors}{hex color codes}
\item{url}{url for artwork info}
}
}
\usage{
artwork
}
\value{
a \link[tibble:tibble-package]{tibble}

Source: \url{http://colorlisa.com/}
}
\description{
A table containing various information about each palette.
}
\keyword{datasets}
