% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_palette}
\alias{lisa_palette}
\title{Call or modify lisa palettes}
\usage{
lisa_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. See \code{names(geopalettes)}}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
Call or modify lisa palettes
}
